Refactor simulation and processing modules
